Q3/TECHIN 515/Lab 4/gesture_capture/requirements.txt
pyserial>=3.5
# Do not install the 'serial' package as it conflicts with pyserial 